backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pdfplumber==0.11.2
python-docx==1.1.2
sentence-transformers==3.0.1
faiss-cpu==1.8.0
numpy==1.26.4
pydantic==2.7.4
orjson==3.10.3

